an implementation of the Android system properties server
ported to KF5 and Qt5. besides running in the target system, it
also gives new compile-time checking to thigs like the signal/slot
connections we didn't have before on Qt4.

from the original commit (since squashed down to just this one):

normally this lives in the kernel, but we configure and build the
shashlik forks of the android libs to use a local server. this is
the same way it is built when run in qemu for doing android app
devel, so this hack is something we ought to be able to rely on.
if Android changes this, all their dev kits break, and i doubt
they will do that :)

the "documentation" for the protocol used can be found in the
Android source code at: android-core/libcutils/properties.c