build libhardware
this is not perfect: it requires access to the NDK API for windows
which is just an API definition and not an implementation.

we are missing an implementation of this.

it also requires a patch to android-core/include/system/ to patch
out use of libsync which is specific to the android kernel, which
provides a kernel implementation of fence based syncs.

so that's two strikes -> linux kernel specifics in an API we
don't have an implementation for.

needs more research